requirements.txt
click==8.2.1
ffmpeg-python==0.2.0
future==1.0.0
numpy==2.2.6
opencv-python==4.12.0.88
platformdirs==4.5.1
scenedetect==0.6.7.1
tqdm==4.67.1
azure-identity==1.17.1
azure-storage-blob==12.22.0
